feat: Item

+++ Source/NBCCH3_5/Public/HealItem.h
@@ -18,7 +18,7 @@ public:
 	AHealItem();
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Item")
-	float HealAmount;
+	int32 HealAmount;
 
 	virtual void ActivateItem(AActor* Activator) override;
 };
